r10781/sofa : Fix minor warnings (unused variables/typedefs; narrowing)
Former-commit-id: 1653b74b4806b97c97bb8e5f732d875a18283d7a

+++ applications/plugins/SofaCUDA/sofa/gpu/cuda/CudaBarycentricMapping-3f1-3f.cpp
@@ -53,7 +53,7 @@ using namespace sofa::gpu::cuda;
 template<>
 void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3fTypes>::apply( Out::VecCoord& out, const In::VecCoord& in )
 {
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     out.fastResize(map.size());
     RegularGridMapperCuda3f1_3f_apply(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 }
@@ -61,7 +61,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3fTypes>::apply
 template<>
 void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3fTypes>::applyJ( Out::VecDeriv& out, const In::VecDeriv& in )
 {
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     out.fastResize(map.size());
     RegularGridMapperCuda3f1_3f_applyJ(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 }
@@ -71,7 +71,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3fTypes>::apply
 {
     calcMapT();
     if (map.size() == 0) return;
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     unsigned int insize = out.size();
 
     RegularGridMapperCuda3f1_3f_applyJT(insize, maxNOut, gridsize, mapT.deviceRead(), out.deviceWrite(), in.deviceRead());
@@ -92,7 +92,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3fTypes>::draw 
 template<>
 void BarycentricMapperRegularGridTopology<CudaVec3fTypes,CudaVec3f1Types>::apply( Out::VecCoord& out, const In::VecCoord& in )
 {
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     out.fastResize(map.size());
     RegularGridMapperCuda3f_3f1_apply(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 }
@@ -100,7 +100,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3fTypes,CudaVec3f1Types>::apply
 template<>
 void BarycentricMapperRegularGridTopology<CudaVec3fTypes,CudaVec3f1Types>::applyJ( Out::VecDeriv& out, const In::VecDeriv& in )
 {
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     out.fastResize(map.size());
     RegularGridMapperCuda3f_3f1_applyJ(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 }
@@ -110,7 +110,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3fTypes,CudaVec3f1Types>::apply
 {
     calcMapT();
     if (map.size() == 0) return;
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     unsigned int insize = out.size();
 
     RegularGridMapperCuda3f_3f1_applyJT(insize, maxNOut, gridsize, mapT.deviceRead(), out.deviceWrite(), in.deviceRead());
